Create PimoroniDiaplayTemperatur4C.py
A great accessory for the Raspberry Pi Pico is the Pico Display from Pimoroni. With this script you can display the temperature in different color combinations. You can change this at the push of a button. This is the temperature sensor, which is installed directly on the Pico Board. You don't need any additional accessories such as sensors.