Create binaeruhr.py
The basis for the binary clock is our LED panel with 25 LEDs in a 5x5 arrangement, i.e. 5 rows and 5 columns. We also use the LED tableau for the letter project with which we can display the alpabeth in individual letters. However, we do not need all the LEDs on the panel for the binary clock. But we still face a problem here. The LED binary clock can only be set at a certain time. The script must be started at this time. Currently we are still missing the solution that the Pico gets the real time and then the script is started at this point at this point. But let's neglect that for now.