Create clock.py
A simple python script to run a clock using the Raspberry Pi Pico to perform script with an IDE like Thonny. The limitation is but the clock must be set to the correct time in the script. Otherwise it is just a timer if you start with s=0, m=0 and h=0. See line 9-12.  If you find  a possibility to import a real time from a timer then let me know. The pico does not have a timer on board and cannot carry out daytim expansion. If you like to change the hour format from 24 to 12 set count3 in range of 12 in line 14.